requirements.txt
pyserial
PyQt5
matplotlib
sqlite3
